requirements.txt
Django==1.5.4
South==0.8.2
argparse==1.2.1
defusedxml==0.4.1
distribute==0.6.24
dj-database-url==0.2.2
dj-static==0.0.5
django-crispy-forms==1.4.0
django-tastypie==0.10.0
django-toolbelt==0.0.1
gunicorn==18.0
lxml==3.2.3
mimeparse==0.1.3
psycopg2==2.5.1
python-dateutil==2.1
six==1.4.1
static==0.4
wsgiref==0.1.2
